pyproject.toml
[project]
name = "crack-tcaptcha"
version = "0.1.0"
description = "Automated TCaptcha (Tencent CAPTCHA) solver — slider & icon-click"
readme = "README.md"
license = "GPL-3.0-or-later"
requires-python = ">=3.10"
authors = [{ name = "gehonglu" }]
keywords = ["captcha", "tcaptcha", "tencent", "slider", "security-research"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Security",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "numpy>=1.24",
    "Pillow>=10.0",
]

[project.optional-dependencies]
icon-click = ["ddddocr>=1.4"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "hypothesis>=6.0",
    "respx>=0.21",
    "ruff>=0.4",
]
docs = [
    "mkdocs-material>=9.0",
]

[project.scripts]
crack-tcaptcha = "crack_tcaptcha.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/crack_tcaptcha"]

[tool.ruff]
line-length = 120
target-version = "py310"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
markers = [
    "network: requires network access (deselect with '-m not network')",
]
asyncio_mode = "auto"
